LLVM Clang test; MD-84
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@10887 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ AYDicomPrint/AYDicomPrintWindowController.mm
@@ -340,7 +340,7 @@ NSString *mediumTag[] = {@"Blue Film", @"Clear Film", @"Paper"};
 		}
 		else
 		{
-			[currentPrinter setObject: [NSString stringWithFormat: @"%d", index-1]  forKey:@"imageDisplayFormatTag"];
+			[currentPrinter setObject: [NSString stringWithFormat: @"%d", (int) index-1]  forKey:@"imageDisplayFormatTag"];
 			ipp = imageDisplayFormatNumbers[[[dict valueForKey: @"imageDisplayFormatTag"] intValue]];
 		}
 	}
